+++ filehandler/EditFileHandler.php
@@ -22,6 +22,11 @@ class EditFileHandler extends BaseFileHandler
     /**
      * @inheritdoc
      */
+    public $position = self::POSITION_TOP;
+
+    /**
+     * @inheritdoc
+     */
     public function getLinkAttributes()
     {
         return [